### 2021-08-24 #76 @ranocha
> Failed to discover a `Statement of need` section in paper

I couldn't find the requirement for a `Statement of need` section in paper in the [author guidelines](https://juliacon.github.io/proceedings-guide/author/) for JuliaCon. It's of course required for JOSS papers. If we need to have this section in our JuliaCon paper, we could basically rename the "Introduction" to "Statement of need".
