### 2021-08-24 #76 @ranocha
The software report is not really helpful since we set up a dedicated repository with reproducibility material for this paper. Our Julia package Trixi.jl is available in [another repository](https://github.com/trixi-framework/Trixi.jl).
