### 2021-08-24 #76 @ranocha
We propose Carsten Bauer (`@crstnbr`) as editor.

Possible reviewers are
- Jason Hicken, Rensselaer Polytechnic Institute (`@jehicken`)
- Lucas C. Wilcox, Naval Postgraduate School (`@lcw`)
- David C. Del Rey Fernández, University of Waterloo (`@dcdelrey`)
- Will Pazner, LLNL (`@pazner` )
- Simon Byrne, Caltech (`@simonbyrne`)
