### 2021-08-24 6086d15
avoid using amcros in captions and section titles
