### 2021-08-24 #76 @ranocha
I can't reproduce the error of whedon locally. Running
```bash
latexmk -c; latexmk -bibtex -pdf paper.tex
```
works fine on my system. Is there a way I can help debugging this? Can we look at `paper.log`?
