### 2021-08-24 #76 @ranocha
This is the PDF generated on my system: [paper.pdf](https://github.com/JuliaCon/proceedings-review/files/7037984/paper.pdf)
