+++ paper/ref.bib
@@ -95,14 +95,19 @@
   note={\url{https://github.com/JuliaParallel/MPI.jl}}
 }
 
-@misc{ranocha2021sbp,
+@article{ranocha2021sbp,
   title={{SummationByPartsOperators.jl}: {A} {J}ulia library of provably stable
          semidiscretization techniques with mimetic properties},
   author={Ranocha, Hendrik},
+  journal={Journal of Open Source Software},
   year={2021},
-  howpublished={\url{https://github.com/ranocha/SummationByPartsOperators.jl}},
-  doi={10.5281/zenodo.4773575},
-  note={Submitted to the Journal of Open Source Software. \url{https://github.com/openjournals/joss-papers/blob/joss.03454/joss.03454/10.21105.joss.03454.pdf}}
+  month={08},
+  doi={10.21105/joss.03454},
+  volume={6},
+  number={64},
+  pages={3454},
+  publisher={The Open Journal},
+  url={https://github.com/ranocha/SummationByPartsOperators.jl}
 }
 
 @online{innes2018fashionable,
