### 2021-09-01 #76 @sloede
I think it was a simple typo:
![image](https://user-images.githubusercontent.com/3637659/131683263-41682b0d-4802-448d-835f-88c7218382d0.png)

Thanks for the assistance everyone! If we only could figure out why `*.aux` needs to be deleted and how to prevent it...
