### 2021-09-01 #76 @arfon
Well, I fixed the first error with https://github.com/openjournals/whedon/commit/c5c16aedb3d6b4377181e859316ed70f1ac0a0b7, but I'm not sure what this new one is about. @matbesancon - the [core code for compiling from TeX](https://github.com/openjournals/whedon/blob/c5c16aedb3d6b4377181e859316ed70f1ac0a0b7/lib/whedon/compilers.rb#L33-L79) hasn't changed for a long time.

I wonder if you could assist here to see if there's something inherently different about this paper? I'm struggling a little to figure out what might be going wrong here...
