### 2021-09-01 #76 @arfon
OK, managed to get an error from the log. Not sure what this means (🤣) but perhaps you all will understand...

```
LaTeX Warning: You have requested, on input line 16, version
               `2020/01/12' of package expl3,
               but only version
               `2018/02/21 L3 programming layer (loader) '
               is available.

(/app/.texlive/texmf-dist/tex/latex/l3packages/xparse/xparse.sty)

! Package siunitx Error: Support package expl3 too old.

See the siunitx package documentation for explanation.
Type  H <return>  for immediate help.
```
