### 2021-09-01 #76 @vchuravy
Hm... Maybe the best thing for now would be to add a `cat paper.log`? In case of failure?

> but perhaps you all will understand...

That looks helpful xD. Can we update the latex installation used?
