### 2021-09-01 #76 @ranocha
I tried to manually revert `siunitx.sty` to an older version from 2011 - we still get the error above
