### 2021-09-01 #76 @ranocha
An alternative to updating the LaTeX installation might be to use a bigger set of standard packages including `siunitx.sty` (bundled in standard texlive distributions).
