### 2021-09-06 3d17342
Allow using pre-generated PDFs from pgfplots
