### 2021-09-06 053367c
Add siunit-*.cfg files
Note: these files (and siunitx.sty) are from siunitx-2.7t.
