### 2021-09-06 #76 @sloede
@carstenbauer after playing some TeX-pong with Whedon I finally managed to have it generate the PDF for our submission :tada:

Please let me know if there is anything else you need from us at the moment. A list of potential reviewers was submitted by @ranocha [above](https://github.com/JuliaCon/proceedings-review/issues/76#issuecomment-904402439).
