### 2021-09-13 #77 @whedon
[REVIEW]: Adaptive numerical simulations with Trixi.jl: A case study of Julia for scientific computing
**Submitting author:** @ranocha (<a href="http://orcid.org/0000-0002-3456-2277">Hendrik Ranocha</a>)
**Repository:** <a href="https://github.com/trixi-framework/paper-2021-juliacon" target ="_blank">https://github.com/trixi-framework/paper-2021-juliacon</a>
**Version:**
