### 2021-09-15 #77 @ranocha
As mentioned in the pre-review issue #76, the software report is not really helpful since we set up a dedicated repository with reproducibility material for this paper. Our Julia package Trixi.jl is available in [another repository](https://github.com/trixi-framework/Trixi.jl).
