### 2021-10-31 #77 @ranocha
Thank you very much for investing your time to help us improving our paper, @simonbyrne. We thought about the location of our paper draft based on the [guide for authors](https://juliacon.github.io/proceedings-guide/author/#submission_details) reproduced below.

![image](https://user-images.githubusercontent.com/12693098/139573155-23cf6396-54da-4589-9091-f2cad3e61949.png)

Since the highlighted note explicitly states

> It is possible to have the paper in a separate repository

we decided to use this option. Our main reason for not putting the paper draft in the Trixi.jl repository is that it would considerably increases the size of our main repo. The paper repo is already ca. 5 MB large and that's without possible iterations during the review. Deleting the branch after publication is also not an option for us, since we specifically want to have a reproducibility repository that is permanent. Also, this does not scale if we want to submit another paper using a similar submission process based on Trixi.jl.

Thus, we would like to keep the paper draft as a separate repo to not clutter Trixi.jl. At the same time, this paper repo is designed as a permanent reproducibility repository like we have for all our papers. We hope that the link to the software repository can be changed by hand upon publication.
