### 2022-01-04 #77 @johnfgibson
@carstenbauer : Has my invitation to review been revoked? When click on https://github.com/openjournals/joss-reviews/invitations , I get the error message "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." I am logged on to my github account.
