### 2022-01-04 trixi-framework/paper-2021-juliacon#20 @johnfgibson
review of manuscript
The authors present Trixi.jl, a Julia package for adaptive high-order simulation of hyperbolic PDEs in 1-3 dimensions, discontinuous Galerkin methods, adaptive meshing, and method-of-lines time-integration. The capabilities and design of Trixi.jl are discussed and illustrated with several two-d quasilinear and nonlinear PDEs with cartesian and complex geometries. Detailed perfomance comparisons to a comparable MPI-based Fortran code run in serial mode are presented for simulations of the incompressible Euler equations and ideal magnetohydrodynamics in a 3d Cartesian cube with periodic boundary conditions. In these serial runs, Trixi.jl is shown to outperform the Fortran code by roughly a factor of two across a wide range of polynomial degree for the spectral elements (3 to 15). The authors provide a frank assessment of the pros, cons, and unknowns of developing and using PDE simulation code in Julia versus the more traditional choices of C, C++, and Fortran.

I think this is an excellent paper that partially addresses a major question: Does Julia hold promise for simplifying scientific software development for large-scale, high-performance supercomputing, without sacrificing efficiency? That the authors wrote a general-purpose hyperbolic PDE simulator in Julia in one person-year attests to the simplifying power of an modern, expressive, high-level language with interoperable packages. The results of the Julia-Fortran comparisons on 3d Euler equations and MHD show that Julia's expressive power does not necessarily come at the cost of a performance overhead. In fact, here we see a well-written Julia code can run *faster* than an equivalent Fortran code, at least in serial. However, as the authors note in the pro/con section, distributed-memory parallelism is more complicated. For production parallel codes, Julia currently relies on, and is limited by, the same MPI libraries used by C, C++, and Fortran.

I have two suggestions for improving the paper. First I think the paper should make it abundantly clear at the outset that Trixi.jl is a serial code, and that the comparisons are to an MPI-HPC code run in serial mode. Nowhere in the paper is it explicitly stated that Trixi.jl is a serial code. The word "serial" only comes up as "serial performance", not to the code itself. For example, from the Introduction "[W]e compare the serial performance with a mature high-performance computing (HPC) Fortran code..." I have to admit the first time I read the paper I assumed that Trixi.jl had distributed-memory parallelism and was just compared to Fortran in serial runs, until I hit the discussion of the difficulties of distributed-memory parallelism in Julia in section 4.2. I have a feeling I'm not alone in pretty much assuming a multi-d PDE simulator is distributed-memory until told otherwise, and that it would do readers a service to be clearer about this up front. I recommend that the Introduction explicitly state that Trixi.jl is a serial code, and that the comparisons are to an MPI-based Fortran code running in serial mode.

Second, I would find it helpful to have a little more information about the spatial domain and discretization for the comparison runs in Section 3.1. These are described as a periodic box [-1]^3, whose interior is warped by a combination of a sine and cosine functions, and four spectral elements in each spatial direction. Does that mean the physical domain of the Euler or MHD problem is still a periodic cube, but the interior grid is a Cartesian product of some irregularly distributed 1d grids? Or that the physical domain of the equations is a shape obtained by mapping a periodic cube through some function involving sines and cosines, with a grid determined by mapping the regular cubic grid through the same function? If the latter it would have to have some pretty specific constraints to preserve the boundaries and match the fluxes across them. If possible, a figure illustrating (a projection of) the mesh and a computed solution (like figs 1,2 and 3) would be nice.

More minor suggestions:

It would be nice to have indication of the size of the computation in terms of numbers of degrees of freedom. For the [-1,1]^3 domain with four elements per dimension, that's a cube divided into 4^3 = 64 subcubes, and in each a cartesian products of (polydeg+1) polynomials in the three directions? For Euler equations with 5 variables (3 velocity components, 1 pressure, 1 energy) and polydegree=15, that's 64 * 16^3 * 5 = 1.3 million degrees of freedom, or about 84 MB of data per Runge-Kutta step. Correct?

In section 3.2, I'm not entirely comfortable with this statement in section 3.2. "We would like to emphasize that Julia code can be at least as fast as code written in these traditional scientific computing languages." It might be wise to change that to "well-written Julia code" or "properly-written". Also, I'm not sure if in these comparisons, the greater speed of Julia is due the MPI overhead of the Fortran code running in serial mode. Maybe you could reference a paper that contains the Julia Benchmarks, which make the same point with more head-to-head comparison of identical serial algorithms.

Lastly a few grammatical/typo suggestions

pg 3 "positivity-preserving limiting"

pg 4 "step-by-step" and "summation-by-parts finite-difference methods"

pg 5 "kinetic-energy preserving"

https://github.com/JuliaCon/proceedings-review/issues/77
