+++ paper/ref.bib
@@ -543,6 +543,19 @@
   howpublished={\url{https://ranocha.de/blog/Optimizing_EC_Trixi}}
 }
 
+@online{ranocha2021efficient,
+  title={Efficient implementation of modern entropy stable and kinetic energy
+         preserving discontinuous {G}alerkin methods for conservation laws},
+  author={Ranocha, Hendrik and Schlottke-Lakemper, Michael and Chan, Jesse and
+          Rueda-Ram\'{i}rez, Andr{\'e}s M and Winters, Andrew R and
+          Hindenlang, Florian and Gassner, Gregor J},
+  year={2021},
+  month={12},
+  eprint={2112.10517},
+  eprinttype={arxiv},
+  eprintclass={cs.MS}
+}
+
 @online{liu2021oscillation,
   title={An Oscillation-free Discontinuous {G}alerkin Method for Hyperbolic
          Systems},
