+++ paper/paper.tex
@@ -446,8 +446,20 @@ hyperbolic PDE in three space dimensions on curvilinear hexahedral meshes.
 For the problem setup we consider a periodic box of the domain $[-1,1]^3$ with four elements in each
 spatial direction. This results in $64 (\text{\lstinline{polydeg}}+1)^3$ degrees
 of freedom (DG nodes) when polynomials of degree \lstinline{polydeg} are used.
-To make the mesh curvilinear, the interior of the box is heavily warped
-by a combination of sine and cosine functions~\cite{chan2019efficient}. To integrate up to the
+To make the mesh curvilinear, the interior of the box is heavily warped by
+a mapping adapted from \cite{chan2019efficient}.
+Specifically, we map Cartesian reference coordinates $(\xi, \eta, \zeta)$ to
+physical coordinates $(x, y, z)$ via the transformation
+\begin{equation}
+\begin{aligned}
+  y &= \eta  + 0.15 \bigl( \cos(1.5 \pi \xi) \cos(0.5 \pi \eta) \cos(0.5 \pi \zeta) \bigr),
+  \\
+  x &= \xi   + 0.15 \bigl( \cos(0.5 \pi \xi) \cos(2 \pi y) \cos(0.5 \pi \zeta) \bigr),
+  \\
+  z &= \zeta + 0.15 \bigl( \cos(0.5 \pi x) \cos(\pi y) \cos(0.5 \pi \zeta) \bigr).
+\end{aligned}
+\end{equation}
+To integrate up to the
 final time $T=1.0$, both codes use the five-stage, fourth-order low-storage explicit
 Runge-Kutta method of Carpenter and Kennedy~\cite{CarpenterKennedy1994}.
 A stable explicit time step is adaptively computed according to the local maximum wave
@@ -460,6 +472,15 @@ a manufactured solution for the 3D compressible Euler equations and Alfv\'{e}n w
 the 3D ideal magnetohydrodynamics (MHD) equations~\cite{gassner2009,altmann2012}. These initial conditions, available in FLUXO and
 \trixi, are typically used to demonstrate the high-order accuracy and convergence properties
 of the frameworks.
+The curved mesh and the initial density for the compressible Euler problem are
+visualized in Figure~\ref{fig:pid_euler_mesh}.
+
+\begin{figure}[!h]
+  \includegraphics[width=\linewidth]{../figures/pid_euler_mesh}
+  \caption{Initial density of the compressible Euler problem on the curved mesh
+           used for the performance benchmarks.}
+  \label{fig:pid_euler_mesh}
+\end{figure}
 
 For these comparisons we examine the performance of FLUXO and \trixi for the weak
 form as well as flux differencing implementations of the DG solver. For the weak form simulations we compute
