### 2022-01-05 trixi-framework/paper-2021-juliacon#20 @ranocha
Thank you very much for your review and your nice comments, @johnfgibson. We are working on a revision based on your comments in #21.
