### 2022-01-05 ce6d860
change caption of mesh plot to indicate that it is a slice
