+++ paper/paper.tex
@@ -482,8 +482,8 @@ visualized in Figure~\ref{fig:pid_euler_mesh}.
 
 \begin{figure}[!h]
   \includegraphics[width=\linewidth]{../figures/pid_euler_mesh}
-  \caption{Initial density of the compressible Euler problem on the curved mesh
-           used for the performance benchmarks.}
+  \caption{Initial density of the compressible Euler problem on a slice of the
+           periodic, curved mesh used for the performance benchmarks.}
   \label{fig:pid_euler_mesh}
 \end{figure}
 
