### 2022-01-05 d9a9072
Revisions based on the review by John F. Gibson (#21)
* mention number of DOFs

* additional performance reference

* add some missing hyphens

* visualization of IC and mesh for PID runs

* comment on parallelization

* change caption of mesh plot to indicate that it is a slice

* suggestions from Andrew

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>
