+++ paper/paper.tex
@@ -169,6 +169,10 @@ following features:
   \end{itemize}
 \end{itemize}
 
+Currently, \trixi provides shared-memory parallelization via multithreading.
+Distributed-memory parallelization with MPI is available for some mesh types;
+increasing the capability of this feature is subject to ongoing work.
+
 Some of the main features of \trixi are demonstrated in the following figures.
 The detailed physical setups and numerical parameters as well as all code
 necessary to reproduce these figures are available in the reproducibility
@@ -178,6 +182,7 @@ Figure~\ref{fig:jet} demonstrates sub-cell entropy-dissipative shock-capturing
 methods with sub-cell positivity-preserving limiters and adaptive mesh refinement applied
 to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillation}.
 \begin{figure}[!h]
+\vspace{-0.5em}
   \begin{subfigure}{0.53\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_density}
     \caption{Density at time $t = 10^{-3}$.}
@@ -287,7 +292,7 @@ in \trixi can be converted into an ODE problem, which can be solved by methods
 from OrdinaryDiffEq.jl. The flexible callback infrastructure of this ODE library
 allows us to provide extended functionality for \trixi without modifying any
 main loop. In particular, various tasks such as input/output operations, adaptive
-mesh refinement (AMR), and positivity preserving limiting are implemented using
+mesh refinement (AMR), and positivity-preserving limiting are implemented using
 callbacks. In case full control over the time step loop is required, e.g., to experiment with new
 features that are not easily realized with the existing API, \trixi also implements its own time
 integration schemes that mimic the interface of OrdinaryDiffEq.jl and that can be used as drop-in
@@ -413,8 +418,8 @@ nature, multiple dispatch, and just-ahead-of-time compilation allow us to implem
 that are strictly necessary. For example, if a user wants to simulate a new
 physics model only on Cartesian grids with smooth solutions, there is no need
 to implement anything for curvilinear coordinates or shock-capturing approaches.
-In addition, it allows a user to extend \trixi step by step with new capabilities.
-For example, there are ongoing efforts to incorporate summation by parts finite
+In addition, it allows a user to extend \trixi step-by-step with new capabilities.
+For example, there are ongoing efforts to incorporate summation-by-parts finite
 difference methods via SummationByPartsOperators.jl \cite{ranocha2021sbp} and
 discontinuous Galerkin methods on simplex elements via StartUpDG.jl\footnote{\url{https://github.com/jlchan/StartUpDG.jl}} in \trixi.
 This is feasible because the modular design of \trixi gives users the flexibility to pick
@@ -444,8 +449,22 @@ Here, we compare the serial performance of \trixi and FLUXO when solving a
 hyperbolic PDE in three space dimensions on curvilinear hexahedral meshes.
 
 For the problem setup we consider a periodic box of the domain $[-1,1]^3$ with four elements in each
-spatial direction. To make the mesh curvilinear, the interior of the box is heavily warped
-by a combination of sine and cosine functions~\cite{chan2019efficient}. To integrate up to the
+spatial direction. This results in $64 (\text{\lstinline{polydeg}}+1)^3$ degrees
+of freedom (DG nodes) per equation when polynomials of degree \lstinline{polydeg} are used.
+To make the mesh curvilinear, the interior of the box is heavily warped by
+a mapping adapted from \cite{chan2019efficient}.
+Specifically, we map Cartesian reference coordinates $(\xi, \eta, \zeta)$ to
[…]
