+++ paper/paper.tex
@@ -233,6 +233,7 @@ curvilinear and unstructured domain. The 2D quadrilateral mesh used in the simul
 (Figure~\ref{fig:ginger_mesh}) was generated with
 HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}}.
 \begin{figure}[!h]
+\vspace{-0.5em}
   \begin{subfigure}{0.475\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
     \caption{Acoustic pressure at time $t = 0$.}
