+++ paper/paper.tex
@@ -196,6 +196,7 @@ to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillatio
            sub-cell entropy-dissipative shock-capturing methods with sub-cell positivity-preserving
            limiters and adaptive mesh refinement for the compressible Euler equations.}
   \label{fig:jet}
+\vspace{-0.5em}
 \end{figure}
 
 Figure~\ref{fig:kelvin_helmholtzs} demonstrates nonlinear stability
