### 2022-01-05 e00517c
reword parallelization feature
Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
