+++ paper/paper.tex
@@ -170,8 +170,8 @@ following features:
 \end{itemize}
 
 Currently, \trixi provides shared-memory parallelization via multithreading.
-Distributed-memory parallelization with MPI is available for some mesh types;
-increasing the capability of this feature is subject to ongoing work.
+Initial parallelization with MPI is available for some mesh types,
+but full support for distributed-memory parallelism is subject to ongoing work.
 
 Some of the main features of \trixi are demonstrated in the following figures.
 The detailed physical setups and numerical parameters as well as all code
