### 2022-01-05 trixi-framework/paper-2021-juliacon#20 @ranocha
@johnfgibson We merged #21 and improved the paper based on your comments above. Please let me know whether the revised version looks good to you.
