### 2022-01-05 trixi-framework/paper-2021-juliacon#20 @johnfgibson
Great! I'm happy with all the revisions and ready to sign off on the review with a :+1: Currently I can't edit the review checklist, if that's necessary to complete the review. I have pinged @carstenbauer about that.
