### 2022-01-06 trixi-framework/paper-2021-juliacon#20 @ranocha
Thanks a lot! In this case, I will close this issue.
