### 2022-01-07 trixi-framework/paper-2021-juliacon#22 @simonbyrne
Review
Trixi.jl is a very impressive piece of software, and I think the paper does a very great job of introducing and highlighting its features as a software package given the page constraints. My main comments:

1. the paper states "Trixi.jl also handles non-conservative PDE terms as in the shallow water equations”. I was bit confused by this at first as Trixi uses the conservative formulation of the SWE (https://trixi-framework.github.io/Trixi.jl/stable/reference-trixi/#Trixi.ShallowWaterEquations2D) as opposed to the so-called "vector invariant" form of the SWE. This appears to be referring to a source term that involves a spatial gradient. Could this be expressed in a slightly different way?

2. I realise such cross-language comparisons are difficult, but given the familiarity of the authors with both codebases, I would have appreciated any insight the authors might have as to the reason for the performance advantage Trixi appears to have over Fluxo (e.g. more cache-friendly memory layout, better use of SIMD or FMA operations, or utilizing different libraries?)
