### 2022-01-07 #77 @simonbyrne
I have submitted my review here: https://github.com/trixi-framework/paper-2021-juliacon/issues/22
