### 2022-01-07 2c8a50e
clarify nonconservative terms
