### 2022-01-07 trixi-framework/paper-2021-juliacon#22 @ranocha
Thank you very much for your review and your kind words, @simonbyrne! I tried to clarify the aspect of nonconservative terms in #23.

A thorough analysis of reasons for these performance differences would require a huge amount of work, basically starting from scratch in each framework and adding pieces step-by-step, monitoring the performance along the way. Thus, I don't think this is suitable for this paper. However, we will be happy to share any conjectures and insights we have on this.
- A significant difference comes from some optimizations we developed in Trixi.jl and ported to FLUXO later. These are described in [a recent preprint](https://arxiv.org/abs/2112.10517).
- Trixi.jl uses slightly different versions of computationally intensive kernels than FLUXO. I expect this to be the next big impact.
- Trixi.jl stores slightly less information in RAM and re-computes some terms, increasing performance in memory-bound parts.
- Some special functions appear to be faster in Julia, e.g., computing `sincos` instead of calling `sin` and `cos` independently in FORTRAN.
