### 2022-01-07 trixi-framework/paper-2021-juliacon#22 @ranocha
@simonbyrne We merged #23 and improved the paper based on your comments above. Please let me know whether the revised version looks good to you.
