+++ paper/paper.tex
@@ -227,6 +227,7 @@ classical Kelvin-Helmholtz flow instability problem.
            entropy-stable methods and adaptive mesh refinement for the
            compressible Euler equations.}
   \label{fig:kelvin_helmholtzs}
+\vspace{-0.5em}
 \end{figure}
 
 Figure~\ref{fig:pressure_waves} shows the approximation
@@ -262,7 +263,7 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
   \label{fig:pressure_waves}
 \end{figure}
 
-
+\clearpage
 \subsection{High-level overview of the code structure}
 
 \trixi is built from the method of lines. Thus, a discretization of \eqref{eq:hcl}
