### 2022-01-07 83c6207
try to adapt layout to work for the old LaTeX version of whedon
