+++ paper/paper.tex
@@ -183,7 +183,7 @@ Figure~\ref{fig:jet} demonstrates sub-cell entropy-dissipative shock-capturing
 methods with sub-cell positivity-preserving limiters and adaptive mesh refinement applied
 to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillation}.
 \begin{figure}[!h]
-\vspace{-0.5em}
+\vspace{-0.7em}
   \begin{subfigure}{0.53\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_density}
     \caption{Density at time $t = 10^{-3}$.}
@@ -197,7 +197,7 @@ to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillatio
            sub-cell entropy-dissipative shock-capturing methods with sub-cell positivity-preserving
            limiters and adaptive mesh refinement for the compressible Euler equations.}
   \label{fig:jet}
-\vspace{-0.5em}
+\vspace{-0.7em}
 \end{figure}
 
 Figure~\ref{fig:kelvin_helmholtzs} demonstrates nonlinear stability
@@ -227,7 +227,7 @@ classical Kelvin-Helmholtz flow instability problem.
            entropy-stable methods and adaptive mesh refinement for the
            compressible Euler equations.}
   \label{fig:kelvin_helmholtzs}
-\vspace{-0.5em}
+\vspace{-0.7em}
 \end{figure}
 
 Figure~\ref{fig:pressure_waves} shows the approximation
@@ -236,7 +236,7 @@ curvilinear and unstructured domain. The 2D quadrilateral mesh used in the simul
 (Figure~\ref{fig:ginger_mesh}) was generated with
 HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}}.
 \begin{figure}[!h]
-\vspace{-0.5em}
+\vspace{-0.7em}
   \begin{subfigure}{0.475\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
     \caption{Acoustic pressure at time $t = 0$.}
@@ -263,7 +263,7 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
   \label{fig:pressure_waves}
 \end{figure}
 
-\clearpage
+
 \subsection{High-level overview of the code structure}
 
 \trixi is built from the method of lines. Thus, a discretization of \eqref{eq:hcl}
