### 2022-01-07 d40d64e
try to adapt layout to work for the old LaTeX version of whedon
