### 2022-01-07 63dc5b9
Revert "try to adapt layout to work for the old LaTeX version of whedon"
This reverts commit d40d64e194b3c206fe96dbe0230a474c69aa2035.
