### 2022-01-07 trixi-framework/paper-2021-juliacon#22 @simonbyrne
I don't think any further analysis is required, but I do think it would be helpful to include a sentence or two in the discussion of section 3 summarizing some of this. Otherwise it is a bit odd to have the benchmark experiment and results which show Trixi is faster than Fluxo, but no discussion as to why that might be the case (other than the comment "Trixi.jl owes its performance optimizations in part to the code introspection and profiling tools available in Julia").

One additional thing I found as well in section 4.1.1:

> A minor exception is due to Julia being built on top of LLVM, which does not perform the same optimizations as other C/C++/Fortran compilers. In particular, LLVM does not use fused multiply-add (FMA) instructions by default, while GCC and Intel compilers do

This isn't quite correct. This specific optimization is usually termed "FMA contraction", and is supported in the LLVM backend, but typically not enabled by default in frontends as it breaks IEEE semantics. For example, Clang has the `-ffp-contract` option, and it is also enabled by "fastmath" annotations in most languages (including Julia). In fact, the `muladd` function and `@muladd` macro are just annotations to enable this optimization locally.
