+++ paper/paper.tex
@@ -761,8 +761,8 @@ As demonstrated in Section~\ref{sec:performance-comparison},
 Julia is not generically slower than traditional high-performance programming
 languages like C, C++, and Fortran. A minor exception is due to
 Julia being built on top of LLVM, which does not perform the same optimizations as other
-C/C++/Fortran compilers. In particular, LLVM does
-not use fused multiply-add (FMA) instructions by default, while GCC and Intel
+C/C++/Fortran compilers. In particular, most LLVM frontends (including Julia) do
+not enable fused multiply-add (FMA) contraction by default, while GCC and Intel
 compilers do. Thus, one must opt-in to these optimizations explicitly.
 Relying on Julia's code manipulation techniques, this can be achieved with the macro
 \lstinline{@muladd}\footnote{\url{https://github.com/SciML/MuladdMacro.jl}}.
