### 2022-01-11 trixi-framework/paper-2021-juliacon#22 @ranocha
Thank you very much for your review, @simonbyrne. I improved the manuscript based on your comments in #24.

> Fortran itself has looser semantics, most compilers enable it by default.

In my experience, this is not only true for Fortran but also for compilers such as GCC and the Intel ones for other languages including C, e.g., [in this example](https://godbolt.org/#g:!((g:!((g:!((h:codeEditor,i:(fontScale:14,fontUsePx:'0',j:1,lang:c%2B%2B,selection:(endColumn:2,endLineNumber:3,positionColumn:1,positionLineNumber:1,selectionStartColumn:2,selectionStartLineNumber:3,startColumn:1,startLineNumber:1),source:'double+foo(double+a,+double+b,+double+c)+%7B%0A++++return+a+*+b+%2B+c%3B%0A%7D%0A'),l:'5',n:'0',o:'C%2B%2B+source+%231',t:'0')),k:49.73108421953307,l:'4',n:'0',o:'',s:0,t:'0'),(g:!((h:compiler,i:(compiler:g111,filters:(b:'0',binary:'1',commentOnly:'0',demangle:'0',directives:'0',execute:'1',intel:'0',libraryCode:'0',trim:'1'),fontScale:14,fontUsePx:'0',j:1,lang:c%2B%2B,libs:!(),options:'-O2+-march%3Dcore-avx2',selection:(endColumn:12,endLineNumber:3,positionColumn:1,positionLineNumber:1,selectionStartColumn:12,selectionStartLineNumber:3,startColumn:1,startLineNumber:1),source:1),l:'5',n:'0',o:'x86-64+gcc+11.1+(Editor+%231,+Compiler+%231)+C%2B%2B',t:'0')),k:50.26891578046694,l:'4',n:'0',o:'',s:0,t:'0')),l:'2',n:'0',o:'',t:'0')),version:4).
