### 2022-01-11 06577e4
make figures a bit smaller to keep page limit
