+++ paper/paper.tex
@@ -184,12 +184,12 @@ methods with sub-cell positivity-preserving limiters and adaptive mesh refinemen
 to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillation}.
 \begin{figure}[!h]
 \vspace{-0.7em}
-  \begin{subfigure}{0.53\linewidth}
+  \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_density}
     \caption{Density at time $t = 10^{-3}$.}
   \end{subfigure}%
   \hspace*{\fill}
-  \begin{subfigure}{0.47\linewidth}
+  \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_mesh}
     \caption{Mesh at time $t = 10^{-3}$.}
   \end{subfigure}%
@@ -204,22 +204,22 @@ Figure~\ref{fig:kelvin_helmholtzs} demonstrates nonlinear stability
 obtained with entropy-stable methods and adaptive mesh refinement applied to a
 classical Kelvin-Helmholtz flow instability problem.
 \begin{figure}[!h]
-  \begin{subfigure}{0.53\linewidth}
+  \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_density_t2}
     \caption{Density at time $t = 2$.}
   \end{subfigure}%
   \hspace*{\fill}
-  \begin{subfigure}{0.47\linewidth}
+  \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_mesh_t2}
     \caption{Mesh at time $t = 2$.}
   \end{subfigure}%
   \\
-  \begin{subfigure}{0.53\linewidth}
+  \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_density_t3}
     \caption{Density at time $t = 3$.}
   \end{subfigure}%
   \hspace*{\fill}
-  \begin{subfigure}{0.47\linewidth}
+  \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_mesh_t3}
     \caption{Mesh at time $t = 3$.}
   \end{subfigure}%
@@ -237,22 +237,22 @@ curvilinear and unstructured domain. The 2D quadrilateral mesh used in the simul
 HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}}.
 \begin{figure}[!h]
 \vspace{-0.7em}
-  \begin{subfigure}{0.475\linewidth}
+  \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
     \caption{Acoustic pressure at time $t = 0$.}
   \end{subfigure}%
   \hspace*{\fill}
-  \begin{subfigure}{0.475\linewidth}
+  \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_08}
     \caption{Acoustic pressure at time $t = 8$.}
   \end{subfigure}%
   \\
-  \begin{subfigure}{0.475\linewidth}
+  \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_16}
     \caption{Acoustic pressure at time $t = 16$.}
   \end{subfigure}%
   \hspace*{\fill}
-  \begin{subfigure}{0.475\linewidth}
+  \begin{subfigure}{0.46\linewidth}
[…]
