+++ paper/paper.tex
@@ -765,7 +765,8 @@ C/C++/Fortran compilers. In particular, most LLVM frontends (including Julia) do
 not enable fused multiply-add (FMA) contraction by default, while GCC and Intel
 compilers do. Thus, one must opt-in to these optimizations explicitly.
 Relying on Julia's code manipulation techniques, this can be achieved with the macro
-\lstinline{@muladd}\footnote{\url{https://github.com/SciML/MuladdMacro.jl}}.
+\lstinline{@muladd}\footnote{\url{https://github.com/SciML/MuladdMacro.jl}},
+which makes it convenient to use Julia's function \lstinline{muladd}\footnote{\url{https://docs.julialang.org/en/v1/base/math/\#Base.muladd}}.
 
 \subsubsection{Julia encourages good software development practices}
 
