### 2022-01-11 4138f40
rewrite FMA contraction part
