### 2022-01-11 7cbbf6c
improve FMA contraction part using suggestions of Simon Byrne
