+++ paper/paper.tex
@@ -760,14 +760,22 @@ in classical programming languages such as C and Fortran are still dominant
 As demonstrated in Section~\ref{sec:performance-comparison},
 Julia is not generically slower than traditional high-performance programming
 languages like C, C++, and Fortran. A minor exception in the context of scientific
-computing relying on floating point operations is that Julia does not enable
-fused multiply-add (FMA) contraction by default, while GCC and Intel compilers
-do\footnote{An interesting coincidence is that not only Julia but most LLVM-based
-frontends tend not to enable FMA contraction by default.}.
-Thus, one must opt-in to these optimizations explicitly.
-Relying on Julia's code manipulation techniques, this can be achieved with the macro
-\lstinline{@muladd}\footnote{\url{https://github.com/SciML/MuladdMacro.jl}},
-which makes it convenient to use Julia's function \lstinline{muladd}\footnote{\url{https://docs.julialang.org/en/v1/base/math/\#Base.muladd}}.
+computing relying on floating point operations is that Julia does not perform
+automatic fused multiply-add (FMA) contraction, that is replacing a multiplication
+followed by an addition with a single FMA instruction, as it is inconsistent
+with its strict floating point semantics. In other compilers, this optimization
+is controlled by a global compiler option (\lstinline{-ffp-contract} in GCC and
+Clang) and often enabled at reasonable optimization levels such as \lstinline{-O2}
+for GCC and Intel compilers\footnote{An interesting coincidence is that not only
+Julia but most LLVM-based frontends tend not to enable FMA contraction by default.}.
+In Julia, FMA contraction is controlled more locally: the
+\lstinline{muladd}\footnote{\url{https://docs.julialang.org/en/v1/base/math/\#Base.muladd}}
+intrinsic function allows the compiler to evaluate a multiply-add using the most
+efficient method. Relying on Julia's code manipulation techniques, the
+\lstinline{@muladd}\footnote{\url{https://github.com/SciML/MuladdMacro.jl}}
+macro makes it convenient to use the \lstinline{muladd} function by
+syntactically rewriting expressions to insert \lstinline{muladd} in appropriate
+locations.
 
 \subsubsection{Julia encourages good software development practices}
 
