+++ paper/paper.tex
@@ -274,7 +274,8 @@ and uses mostly Runge-Kutta methods implemented in OrdinaryDiffEq.jl, which is p
 of DifferentialEquations.jl \cite{rackauckas2017differentialequations}.
 
 \begin{figure*}[htbp]
-  \includegraphics[width=\linewidth]{../figures/trixi_global_overview}
+\centering
+  \includegraphics[width=0.9\linewidth]{../figures/trixi_global_overview}
   \caption{Schematic overview of the basic components in \trixi and how they
            interact.}
   \label{fig:trixi_global_overview}
@@ -485,7 +486,8 @@ The curved mesh and the initial density for the compressible Euler problem are
 visualized in Figure~\ref{fig:pid_euler_mesh}.
 
 \begin{figure}[!h]
-  \includegraphics[width=\linewidth]{../figures/pid_euler_mesh}
+\centering
+  \includegraphics[width=0.9\linewidth]{../figures/pid_euler_mesh}
   \caption{Initial density of the compressible Euler problem on a slice of the
            periodic, curved mesh used for the performance benchmarks.}
   \label{fig:pid_euler_mesh}
