### 2022-01-11 6411879
Apply suggestions from code review
Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
