+++ paper/paper.tex
@@ -720,8 +720,8 @@ profiling tools available in Julia. Similar tools used to optimize the performan
 languages are often not as easy to use as their Julia counterparts. Therefore,
 we developed some additional performance improvements in \trixi and ported them
 to FLUXO later \cite{ranocha2021efficient}. Moreover, \trixi uses slightly
-different implementations of computationally intensive kernels and stores a bit
-less information in RAM, recomputing some terms instead. Additionally, the
+different implementations of computationally intensive kernels and stores marginally
+less information in memory, recomputing some terms instead. Additionally, the
 manufactured solution of the compressible Euler equations uses source terms
 containing sine and cosine terms, which are computed together by \lstinline{sincos}
 in \trixi but require individual calls to \lstinline{sin} and \lstinline{cos} in
@@ -763,7 +763,7 @@ As demonstrated in Section~\ref{sec:performance-comparison},
 Julia is not generically slower than traditional high-performance programming
 languages like C, C++, and Fortran. A minor exception in the context of scientific
 computing relying on floating point operations is that Julia does not perform
-automatic fused multiply-add (FMA) contraction, that is replacing a multiplication
+automatic fused multiply-add (FMA) contraction, that is, replacing a multiplication
 followed by an addition with a single FMA instruction, as it is inconsistent
 with its strict floating point semantics. In other compilers, this optimization
 is controlled by a global compiler option (\lstinline{-ffp-contract} in GCC and
