### 2022-01-11 f9efbe4
move FMA discussion to the appendix
