### 2022-01-11 0a18e33
review by Simon Byrne (#24)
* some comments on performance difference

* reword FMA part

* crop figure of PID test mesh to keep page limit

* make figures a bit smaller to keep page limit

* link to Base muladd

* rewrite FMA contraction part

* improve FMA contraction part using suggestions of Simon Byrne

* tune figures to keep page limit

* Apply suggestions from code review

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>

* move FMA discussion to the appendix

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
