### 2022-01-11 trixi-framework/paper-2021-juliacon#22 @ranocha
Thanks again for your constructive criticism, @simonbyrne. We merged PR #24 improving the manuscript based on your comments. Please let me know whether everything looks good to you.
