### 2022-01-11 #77 @ranocha
Thanks again to the reviewers. We improved the manuscript based on the suggestions of @johnfgibson and @simonbyrne.
The new manuscript built by whedon hits 11 pages. As described above, I get only 10 pages locally, see [paper.pdf](https://github.com/JuliaCon/proceedings-review/files/7846828/paper.pdf). It looks like some float placement logic changed from the (rather old) LaTeX installation of whedon to the (newer) version I use locally:
```bash
$ pdflatex --version
pdfTeX 3.14159265-2.6-1.40.20 (TeX Live 2019/Debian)
kpathsea version 6.3.1
Copyright 2019 Han The Thanh (pdfTeX) et al.
There is NO warranty.  Redistribution of this software is
covered by the terms of both the pdfTeX copyright and
the Lesser GNU General Public License.
For more information about these matters, see the file
named COPYING and the pdfTeX source.
Primary author of pdfTeX: Han The Thanh (pdfTeX) et al.
Compiled with libpng 1.6.37; using libpng 1.6.37
Compiled with zlib 1.2.11; using zlib 1.2.11
Compiled with xpdf version 4.01
```
