### 2022-01-12 #77 @ranocha
@carstenbauer What's your opinion on the page limit issue?
- Is it okay to have a few lines on page 11 (although that looks bad)?
- Is there any chance to update the LaTeX installation you use for this journal?
- Or do we need to find a way to convince whedon to generate the same layout I get locally? If so, is there any better way to do this than pushing to our repo and asking whedon to generate a new PDF here?
