### 2022-01-12 #77 @simonbyrne
I have completed my review as well.

I haven't personally run the performance benchmarks either, but I did verify the code to reproduce the results is in the paper repository, and that their methodology seems appropriate.
