### 2022-01-13 trixi-framework/paper-2021-juliacon#22 @ranocha
@simonbyrne Can we close this issue or is there anything else you would like to see changed?
