### 2022-01-13 trixi-framework/paper-2021-juliacon#22 @simonbyrne
No, this can be closed.

Thank you for the paper, it was a great read.
