### 2022-01-13 trixi-framework/paper-2021-juliacon#22 @ranocha
Well, thank you for your nice review and your help in improving our manuscript.
