+++ paper/paper.tex
@@ -261,6 +261,7 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
                 perturbation equations at three points in time as well as the unstructured,
                 curvilinear quadrilateral mesh.}
   \label{fig:pressure_waves}
+\vspace{-0.7em}
 \end{figure}
 
 
@@ -273,7 +274,7 @@ integration method. Currently, \trixi focuses on the spatial semidiscretization
 and uses mostly Runge-Kutta methods implemented in OrdinaryDiffEq.jl, which is part
 of DifferentialEquations.jl \cite{rackauckas2017differentialequations}.
 
-\begin{figure*}[htbp]
+\begin{figure*}[th]
 \centering
   \includegraphics[width=0.9\linewidth]{../figures/trixi_global_overview}
   \caption{Schematic overview of the basic components in \trixi and how they
