+++ paper/paper.tex
@@ -19,6 +19,9 @@
 % Use multiple images in a single environment
 \usepackage{subcaption}
 
+% Option [H] for figure placement
+\usepackage{float}
+
 % Allow external generation of images to speed up compilation
 \usepackage{tikz}
 \usetikzlibrary{external}
@@ -182,7 +185,7 @@ repository for this article \cite{ranocha2021adaptiveRepro}.
 Figure~\ref{fig:jet} demonstrates sub-cell entropy-dissipative shock-capturing
 methods with sub-cell positivity-preserving limiters and adaptive mesh refinement applied
 to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillation}.
-\begin{figure}[!h]
+\begin{figure}[!ht]
 \vspace{-0.7em}
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_density}
@@ -203,7 +206,7 @@ to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillatio
 Figure~\ref{fig:kelvin_helmholtzs} demonstrates nonlinear stability
 obtained with entropy-stable methods and adaptive mesh refinement applied to a
 classical Kelvin-Helmholtz flow instability problem.
-\begin{figure}[!h]
+\begin{figure}[!ht]
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_density_t2}
     \caption{Density at time $t = 2$.}
@@ -235,7 +238,7 @@ of acoustic perturbation equations~\cite{ewert2003acoustic} wave scattering on a
 curvilinear and unstructured domain. The 2D quadrilateral mesh used in the simulation
 (Figure~\ref{fig:ginger_mesh}) was generated with
 HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}}.
-\begin{figure}[!h]
+\begin{figure}[H]
 \vspace{-0.7em}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
