+++ paper/paper.tex
@@ -282,6 +282,7 @@ of DifferentialEquations.jl \cite{rackauckas2017differentialequations}.
   \caption{Schematic overview of the basic components in \trixi and how they
            interact.}
   \label{fig:trixi_global_overview}
+\vspace{-0.7em}
 \end{figure*}
 
 Figure~\ref{fig:trixi_global_overview} presents an overview of the basic
