+++ paper/paper.tex
@@ -48,6 +48,8 @@
 
 \newcommand{\todo}[1]{{\color{red}#1}}
 
+\sloppy
+
 
 \begin{document}
 
