+++ paper/paper.tex
@@ -48,8 +48,6 @@
 
 \newcommand{\todo}[1]{{\color{red}#1}}
 
-\sloppy
-
 
 \begin{document}
 
