+++ paper/paper.tex
@@ -197,7 +197,7 @@ to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillatio
            sub-cell entropy-dissipative shock-capturing methods with sub-cell positivity-preserving
            limiters and adaptive mesh refinement for the compressible Euler equations.}
   \label{fig:jet}
-\vspace{-0.7em}
+\vspace{-1.4em}
 \end{figure}
 
 Figure~\ref{fig:kelvin_helmholtzs} demonstrates nonlinear stability
@@ -210,7 +210,7 @@ curvilinear and unstructured domain. The 2D quadrilateral mesh used in the simul
 (Figure~\ref{fig:ginger_mesh}) was generated with
 HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}}.
 
-\begin{figure}[!h]
+\begin{figure}[!ht]
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_density_t2}
     \caption{Density at time $t = 2$.}
@@ -237,7 +237,7 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
 \vspace{-0.7em}
 \end{figure}
 %
-\begin{figure}[!h]
+\begin{figure}[!ht]
 \vspace{-0.7em}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
