### 2022-01-14 e7b6579
\captionsetup[subfigure]{font=footnotesize}
