+++ paper/paper.tex
@@ -183,6 +183,7 @@ Figure~\ref{fig:jet} demonstrates sub-cell entropy-dissipative shock-capturing
 methods with sub-cell positivity-preserving limiters and adaptive mesh refinement applied
 to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillation}.
 \begin{figure}[!h]
+\captionsetup[subfigure]{font=footnotesize}
 \vspace{-0.7em}
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_density}
@@ -211,6 +212,7 @@ curvilinear and unstructured domain. The 2D quadrilateral mesh used in the simul
 HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}}.
 
 \begin{figure}[!ht]
+\captionsetup[subfigure]{font=footnotesize}
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_density_t2}
     \caption{Density at time $t = 2$.}
@@ -238,6 +240,7 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
 \end{figure}
 %
 \begin{figure}[!ht]
+\captionsetup[subfigure]{font=footnotesize}
 \vspace{-0.7em}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
@@ -544,6 +547,7 @@ a relative comparison using FLUXO as reference in the bottom portion; smaller va
 better.
 
 \begin{figure}[!ht]
+\captionsetup[subfigure]{font=footnotesize}
 \centering
   \begin{subfigure}{\linewidth}
     \tikzsetnextfilename{runtime_euler_absolute}
@@ -623,6 +627,7 @@ better.
 \end{figure}
 
 \begin{figure}[!ht]
+\captionsetup[subfigure]{font=footnotesize}
 \centering
   \begin{subfigure}{\linewidth}
     \tikzsetnextfilename{runtime_mhd_absolute}
