### 2022-01-14 9b57bfd
Revert "\captionsetup[subfigure]{font=footnotesize}"
This reverts commit e7b6579a2896b28f20d8ed721f0b4bbf15b9cc78.
