+++ paper/paper.tex
@@ -241,17 +241,17 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
 \vspace{-0.7em}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
-    \caption{Acoustic pressure at time $t = 0$.}
+    \caption{Acoustic pressure at $t = 0$.}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_08}
-    \caption{Acoustic pressure at time $t = 8$.}
+    \caption{Acoustic pressure at $t = 8$.}
   \end{subfigure}%
   \\
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_16}
-    \caption{Acoustic pressure at time $t = 16$.}
+    \caption{Acoustic pressure at $t = 16$.}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
@@ -260,8 +260,8 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
     \label{fig:ginger_mesh}
   \end{subfigure}%
   \caption{Numerical solutions of pressure wave scattering for the acoustic
-                perturbation equations at three points in time as well as the unstructured,
-                curvilinear quadrilateral mesh.}
+                perturbation equations at three points in time $t$ as well as
+                the unstructured, curvilinear quadrilateral mesh.}
   \label{fig:pressure_waves}
 \vspace{-0.7em}
 \end{figure}
