### 2022-01-14 a708d87
shorten subcaptions of fig:pressure_waves even more
