+++ paper/paper.tex
@@ -241,17 +241,17 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
 \vspace{-0.7em}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
-    \caption{Acoustic pressure at $t = 0$.}
+    \caption{Acoustic pressure, $t = 0$.}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_08}
-    \caption{Acoustic pressure at $t = 8$.}
+    \caption{Acoustic pressure, $t = 8$.}
   \end{subfigure}%
   \\
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_16}
-    \caption{Acoustic pressure at $t = 16$.}
+    \caption{Acoustic pressure, $t = 16$.}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
