### 2022-01-14 26182ff
set fontsize of subcaptions explicitly
