+++ paper/paper.tex
@@ -186,12 +186,12 @@ to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillatio
 \vspace{-0.7em}
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_density}
-    \caption{Density at time $t = 10^{-3}$.}
+    \caption{{\fontsize{7}{10}\selectfont Density at time $t = 10^{-3}$.}}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/jet_mesh}
-    \caption{Mesh at time $t = 10^{-3}$.}
+    \caption{{\fontsize{7}{10}\selectfont Mesh at time $t = 10^{-3}$.}}
   \end{subfigure}%
   \caption{Numerical solutions of a supersonic jet with Mach number 2000 using
            sub-cell entropy-dissipative shock-capturing methods with sub-cell positivity-preserving
@@ -213,22 +213,22 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
 \begin{figure}[!ht]
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_density_t2}
-    \caption{Density at time $t = 2$.}
+    \caption{{\fontsize{7}{10}\selectfont Density at time $t = 2$.}}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_mesh_t2}
-    \caption{Mesh at time $t = 2$.}
+    \caption{{\fontsize{7}{10}\selectfont Mesh at time $t = 2$.}}
   \end{subfigure}%
   \\
   \begin{subfigure}{0.52\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_density_t3}
-    \caption{Density at time $t = 3$.}
+    \caption{{\fontsize{7}{10}\selectfont Density at time $t = 3$.}}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/kelvin_helmholtz_mesh_t3}
-    \caption{Mesh at time $t = 3$.}
+    \caption{{\fontsize{7}{10}\selectfont Mesh at time $t = 3$.}}
   \end{subfigure}%
   \caption{Numerical solutions of a Kelvin-Helmholtz instability using
            entropy-stable methods and adaptive mesh refinement for the
@@ -241,26 +241,26 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
 \vspace{-0.7em}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
-    \caption{Acoustic pressure, $t = 0$.}
+    \caption{{\fontsize{7}{10}\selectfont Acoustic pressure at time $t = 0$.}}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_08}
-    \caption{Acoustic pressure, $t = 8$.}
+    \caption{{\fontsize{7}{10}\selectfont Acoustic pressure at time $t = 8$.}}
   \end{subfigure}%
   \\
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_16}
-    \caption{Acoustic pressure, $t = 16$.}
[…]
