### 2022-01-14 6362e93
Revert "set fontsize of subcaptions explicitly"
This reverts commit 26182ff366e8fc036c42ae1c2e3ec227dd115fb2.
