+++ paper/paper.tex
@@ -17,7 +17,7 @@
 % \sisetup{tight-spacing=true}
 
 % Use multiple images in a single environment
-\usepackage{subcaption}
+\usepackage[font=footnotesize]{subcaption}
 
 % Allow external generation of images to speed up compilation
 \usepackage{tikz}
