+++ paper/paper.tex
@@ -17,7 +17,8 @@
 % \sisetup{tight-spacing=true}
 
 % Use multiple images in a single environment
-\usepackage[font=scriptsize]{subcaption}
+\usepackage[font=scriptsize]{subcaption} % required for the old LaTeX installation of whedon
+% \usepackage{subcaption} % required for newer LaTeX installations
 
 % Allow external generation of images to speed up compilation
 \usepackage{tikz}
@@ -197,7 +198,7 @@ to an astrophysical supersonic jet with Mach number 2000 \cite{liu2021oscillatio
            sub-cell entropy-dissipative shock-capturing methods with sub-cell positivity-preserving
            limiters and adaptive mesh refinement for the compressible Euler equations.}
   \label{fig:jet}
-\vspace{-1.4em}
+\vspace{-0.7em}
 \end{figure}
 
 Figure~\ref{fig:kelvin_helmholtzs} demonstrates nonlinear stability
@@ -241,17 +242,17 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
 \vspace{-0.7em}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_00}
-    \caption{Acoustic pressure, $t = 0$.}
+    \caption{Acoustic pressure at time $t = 0$.}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_08}
-    \caption{Acoustic pressure, $t = 8$.}
+    \caption{Acoustic pressure at time $t = 8$.}
   \end{subfigure}%
   \\
   \begin{subfigure}{0.46\linewidth}
     \includegraphics[width=\textwidth]{../figures/ginger_time_16}
-    \caption{Acoustic pressure, $t = 16$.}
+    \caption{Acoustic pressure at time $t = 16$.}
   \end{subfigure}%
   \hspace*{\fill}
   \begin{subfigure}{0.46\linewidth}
@@ -260,7 +261,7 @@ HOHQMesh.jl\xspace\footnote{\url{https://github.com/trixi-framework/HOHQMesh.jl}
     \label{fig:ginger_mesh}
   \end{subfigure}%
   \caption{Numerical solutions of pressure wave scattering for the acoustic
-                perturbation equations at three points in time $t$ as well as
+                perturbation equations at three points in time as well as
                 the unstructured, curvilinear quadrilateral mesh.}
   \label{fig:pressure_waves}
 \vspace{-0.7em}
