### 2022-01-14 e3904c6
convince whedon (#25)
* WIP

* try H from package float

* Revert "try H from package float"

This reverts commit 181570596d2fa048598a8e27c7abd9ac7566a427.

* reorder texts and figures in the source

* WIP

* WIP

* try sloppy

* Revert "try sloppy"

This reverts commit c52c3701e47db871551698f1290e5081bda03277.

* WIP

* \captionsetup[subfigure]{font=footnotesize}

* Revert "\captionsetup[subfigure]{font=footnotesize}"

This reverts commit e7b6579a2896b28f20d8ed721f0b4bbf15b9cc78.

* shorten subcaptions of fig:pressure_waves

* shorten subcaptions of fig:pressure_waves even more

* set fontsiz
[…]
