### 2022-01-14 #77 @ranocha
After two hours of annoying debugging, I found and fixed the issue:
The old LaTeX installation of whedon has a bug in the subcaption package/setup. The labels of subfigures are too big by default, even bigger than the main captions of figures.
See https://github.com/trixi-framework/paper-2021-juliacon/pull/25 for details.

@carstenbauer The current draft complies with the page limit (10 pages). Is there anything else we should do?
