+++ paper/ref.bib
@@ -269,16 +269,18 @@
   year={2013}
 }
 
-@online{ranocha2021optimized,
-  title={{Optimized Runge-Kutta Methods with Automatic Step Size Control
-         for Compressible Computational Fluid Dynamics}},
+@article{ranocha2021optimized,
+  title={Optimized Runge-Kutta Methods with Automatic Step Size Control
+         for Compressible Computational Fluid Dynamics},
   author={Ranocha, Hendrik and Dalcin, Lisandro and Parsani, Matteo
           and Ketcheson, David I.},
+  journal={Communications on Applied Mathematics and Computation},
   year={2021},
+  month={11},
+  doi={10.1007/s42967-021-00159-w},
   eprint={2104.06836},
   eprinttype={arxiv},
-  eprintclass={math.NA},
-  note={Accepted in Communications on Applied Mathematics and Computation}
+  eprintclass={math.NA}
 }
 
 @article{besard2018juliagpu,
