### 2022-01-14 #77 @ranocha
Fixed, thanks for the ping (the paper was published in the meantime, we cited only the preprint before)
