+++ paper/ref.bib
@@ -270,7 +270,7 @@
 }
 
 @article{ranocha2021optimized,
-  title={Optimized Runge-Kutta Methods with Automatic Step Size Control
+  title={Optimized {R}unge-{K}utta Methods with Automatic Step Size Control
          for Compressible Computational Fluid Dynamics},
   author={Ranocha, Hendrik and Dalcin, Lisandro and Parsani, Matteo
           and Ketcheson, David I.},
